Prl Created a base hearing actual summary component
