+++ src/hearings/containers/hearing-actuals/hearing-actual-edit-summary/hearing-actual-summary.component.ts
@@ -1,126 +1,44 @@
-import {Component, OnDestroy, OnInit} from '@angular/core';
-import {ActivatedRoute} from '@angular/router';
+import {Component} from '@angular/core';
+import {ActivatedRoute, Router} from '@angular/router';
 import {Store} from '@ngrx/store';
-import * as moment from 'moment';
-import {combineLatest, Observable, Subscription} from 'rxjs';
-import {filter, map, take} from 'rxjs/operators';
 import {
   ActualHearingDayModel,
   HearingActualsMainModel,
-  HearingOutcomeModel
 } from '../../../models/hearingActualsMainModel';
 import {
-  ACTION,
-  AnswerSource,
   HearingActualAddEditSummaryEnum,
-  HearingChannelEnum,
-  HearingDateEnum,
   HearingResult
 } from '../../../models/hearings.enum';
-import {LovRefDataModel} from '../../../models/lovRefData.model';
 import {HearingsService} from '../../../services/hearings.service';
 import * as fromHearingStore from '../../../store';
 import {ActualHearingsUtils} from '../../../utils/actual-hearings.utils';
+import { HearingActualSummaryBaseComponent } from '../hearing-actual-summary-base/hearing-actual-summary-base.component';
 
 @Component({
   selector: 'exui-hearing-actual-summary',
   templateUrl: './hearing-actual-summary.component.html',
   styleUrls: ['./hearing-actual-summary.component.scss']
 })
-export class HearingActualSummaryComponent implements OnInit, OnDestroy {
-  public hearingState$: Observable<fromHearingStore.State>;
-  public isPaperHearing: boolean;
-  public hearingActualsMainModel: HearingActualsMainModel;
-  public hearingOutcome: HearingOutcomeModel;
-  public hearingRoles: LovRefDataModel[] = [];
-  public actualHearingDays: ActualHearingDayModel[];
-  public hearingTypes: LovRefDataModel[];
-  public actualPartHeardReasonCodes: LovRefDataModel[];
-  public actualCancellationReasonCodes: LovRefDataModel[];
-  public hearingResult: string;
-  public hearingTypeDescription: string;
-  public hearingResultReasonTypeDescription: string;
-  public validationErrors: { id: string, message: string }[] = [];
+export class HearingActualSummaryComponent extends HearingActualSummaryBaseComponent {
   public serverErrors: { id: string, message: string }[] = [
     {id: 'serverError', message: 'There was a system error and your request could not be processed. Please try again.'}
   ];
-  public hearingStageResultErrorMessage = '';
   public hearingTimingResultErrorMessage = '';
   public hearingPartiesResultErrorMessage = '';
   public hearingDaysRequiredErrorMessage = '';
-  public successBanner: boolean = false;
-  public submitted = false;
-  public sub: Subscription;
-  public id: string;
-  public errors$: Observable<number>;
-  public partyChannels: LovRefDataModel[] = [];
-  public hearingDateRange: string;
-  public hearingDatesAccordion = {} as { [hearingDate: string]: boolean };
-  public actualHearingUtils = ActualHearingsUtils;
-  public answerSource = AnswerSource;
-  public hearingRequestID: string;
 
-  constructor(private readonly hearingStore: Store<fromHearingStore.State>, private readonly hearingsService: HearingsService, private readonly route: ActivatedRoute) {
-    this.hearingRoles = this.route.snapshot.data.hearingRole;
-    this.hearingTypes = this.route.snapshot.data.hearingTypes;
+  constructor(
+    public readonly hearingStore: Store<fromHearingStore.State>,
+    public readonly hearingsService: HearingsService,
+    public readonly route: ActivatedRoute,
+    public readonly router: Router
+  ) {
+    super(hearingStore, hearingsService, route, router);
     this.partyChannels = this.route.snapshot.data.partyChannel;
-    this.actualPartHeardReasonCodes = this.route.snapshot.data.actualPartHeardReasonCodes;
-    this.actualCancellationReasonCodes = this.route.snapshot.data.cancelHearingActualReasons;
-  }
-
-  public ngOnInit(): void {
-    this.id = this.route.snapshot.params.id;
-    this.hearingState$ = this.hearingStore.select(fromHearingStore.getHearingsFeatureState);
-    this.hearingState$.subscribe(state => {
-      this.isPaperHearing = state.hearingRequest.hearingRequestMainModel.hearingDetails.hearingChannels.includes(HearingChannelEnum.ONPPR);
-    });
-
-    this.errors$ = combineLatest([
-      this.hearingStore.select(fromHearingStore.getHearingActualsLastError),
-      this.hearingStore.select(fromHearingStore.getHearingRequestLastError)
-    ]).pipe(map((errors) => errors.filter(item => item).length));
-
-    this.sub = this.hearingState$.pipe(
-      filter((state) => !!state.hearingActuals.hearingActualsMainModel),
-    )
-      .subscribe((state) => {
-        this.hearingActualsMainModel = state.hearingActuals.hearingActualsMainModel;
-        this.hearingOutcome = this.hearingActualsMainModel.hearingActuals && this.hearingActualsMainModel.hearingActuals.hearingOutcome;
-        this.hearingTypeDescription = this.hearingOutcome && this.hearingOutcome.hearingType && this.getHearingTypeDescription(this.hearingOutcome.hearingType);
-        this.hearingResultReasonTypeDescription = this.hearingOutcome && this.getHearingResultReasonTypeDescription(this.hearingOutcome);
-        this.hearingRequestID = state.hearingRequest.hearingRequestMainModel.requestDetails.hearingRequestID;
-        this.hearingResult = this.hearingOutcome && this.hearingOutcome.hearingResult;
-        this.actualHearingDays = ActualHearingsUtils.getActualHearingDays(this.hearingActualsMainModel);
-        this.hearingDateRange = this.calculateEarliestHearingDate(this.actualHearingDays);
-
-        this.hearingActualsMainModel.hearingPlanned.plannedHearingDays.forEach(
-          plannedDay => {
-            const key = ActualHearingsUtils.getDate(plannedDay.plannedStartTime);
-            if (!this.hearingDatesAccordion.hasOwnProperty(key)) {
-              this.hearingDatesAccordion[key] = false;
-            }
-          });
-      });
-
-
   }
 
   public hearingActualAddEditUrl(): string {
     return `/hearings/actuals/${this.hearingRequestID}/hearing-actual-add-edit-summary`;
-  }
-
-  public ngOnDestroy(): void {
-    if (this.sub) {
-      this.sub.unsubscribe();
-    }
-  }
-
-  public onBack(): void {
-    this.hearingsService.navigateAction(ACTION.BACK);
-  }
-
-  public getHearingDateText(): string {
-    return this.actualHearingDays && this.actualHearingDays.length > 1 ? 'Hearing date(s)' : 'Hearing date';
   }
 
   public onSubmitHearingDetails(): void {
@@ -131,72 +49,12 @@ export class HearingActualSummaryComponent implements OnInit, OnDestroy {
     }
   }
 
-  public getHearingResultReasonTypeDescription(hearingOutcome: HearingOutcomeModel): string {
-    const hearingActualReasonsRefData = hearingOutcome.hearingResult === HearingResult.COMPLETED
-      ? [] : hearingOutcome.hearingResult === HearingResult.ADJOURNED
-        ? this.actualPartHeardReasonCodes : this.actualCancellationReasonCodes;
-
-    const hearingActualReason = hearingActualReasonsRefData && hearingActualReasonsRefData.find(refData => refData.key === hearingOutcome.hearingResultReasonType);
-    if (hearingActualReason) {
-      return hearingActualReason.value_en;
-    }
-    return '';
-  }
-
-  public isHearingAllRequiredDaysCovered(): boolean {
-    let isAllDaysCovered: boolean = true;
-    const isActualHearingDaysAvailable = this.hearingActualsMainModel.hearingActuals && this.hearingActualsMainModel.hearingActuals.actualHearingDays
-      && this.hearingActualsMainModel.hearingActuals.actualHearingDays.length > 0;
-    if (isActualHearingDaysAvailable) {
-      this.actualHearingDays.forEach((actualHearingDay) => {
-        if (!actualHearingDay.notRequired && !this.isDetailsProvidedForDay(actualHearingDay)) {
-          isAllDaysCovered = false;
-        }
-      });
-    }
-    return isAllDaysCovered;
-  }
-
   private isAllHearingActualsTimingAvailable(hearingActualsMainModel: HearingActualsMainModel) {
     const hasAllActualDays = hearingActualsMainModel.hearingActuals && hearingActualsMainModel.hearingActuals.actualHearingDays
       && hearingActualsMainModel.hearingActuals.actualHearingDays.length === hearingActualsMainModel.hearingPlanned.plannedHearingDays.length;
 
     return hasAllActualDays && hearingActualsMainModel.hearingActuals.actualHearingDays.every(
       actualDay => this.isAcutalTimingAvailable(actualDay)
-    );
-  }
-
-  public isAcutalTimingAvailable(actualDay: ActualHearingDayModel): boolean {
-    return actualDay.notRequired || Boolean(actualDay.hearingDate && actualDay.hearingStartTime
-      && actualDay.hearingEndTime && actualDay.pauseDateTimes);
-  }
-
-  public getHearingTypeDescription(hearingType: string): string {
-    const hearingTypeFromLookup = this.hearingTypes && this.hearingTypes.find(x => x.key.toLowerCase() === hearingType.toLowerCase());
-
-    return hearingTypeFromLookup ? hearingTypeFromLookup.value_en : '';
-  }
-
-  public isHearingActualsDaysAvailable(hearingDate: string) {
-    const hearingInfo = this.hearingActualsMainModel.hearingActuals  && this.hearingActualsMainModel.hearingActuals.actualHearingDays
-      && this.hearingActualsMainModel.hearingActuals.actualHearingDays.length
-      && this.hearingActualsMainModel.hearingActuals.actualHearingDays
-        .find((hearingsInfo: ActualHearingDayModel) => hearingsInfo.hearingDate === hearingDate);
-    return !!hearingInfo && this.isAcutalTimingAvailable(hearingInfo);
-  }
-
-  public isHearingActualsPartiesAvailable(hearingDate: string): boolean {
-    const hearingInfo = this.hearingActualsMainModel.hearingActuals && this.hearingActualsMainModel.hearingActuals.actualHearingDays
-      && this.hearingActualsMainModel.hearingActuals.actualHearingDays.length
-      && this.hearingActualsMainModel.hearingActuals.actualHearingDays.find((hearingsInfo: ActualHearingDayModel) => hearingsInfo.hearingDate === hearingDate);
-    return hearingInfo && (hearingInfo.notRequired || hearingInfo.actualDayParties.length > 0);
-  }
-
-  public isAllHearingActualsPartiesAvailable(hearingActualsMainModel: HearingActualsMainModel) {
-    const hasAllActualDays = hearingActualsMainModel.hearingActuals && hearingActualsMainModel.hearingActuals.actualHearingDays
-      && hearingActualsMainModel.hearingActuals.actualHearingDays.length === hearingActualsMainModel.hearingPlanned.plannedHearingDays.length;
-    return hasAllActualDays && hearingActualsMainModel.hearingActuals.actualHearingDays.every(
-      actualDay => actualDay.notRequired || actualDay.actualDayParties.length > 0
     );
   }
 
@@ -233,17 +91,6 @@ export class HearingActualSummaryComponent implements OnInit, OnDestroy {
     } as ActualHearingDayModel;
     const patchedHearingActuals = ActualHearingsUtils.mergeSingleHearingPartActuals
     (this.hearingActualsMainModel, hearingDay.hearingDate, updatedActuals);
-
-    this.hearingStore.dispatch(new fromHearingStore.UpdateHearingActuals({
-      hearingId: this.id,
-      hearingActuals: patchedHearingActuals,
-    }));
-  }
-
-  public changeWasThisHearingDayRequired(hearingDay: ActualHearingDayModel) {
-    this.validationErrors = [];
-    const patchedHearingActuals = ActualHearingsUtils.mergeSingleHearingPartActuals
-    (this.hearingActualsMainModel, hearingDay.hearingDate, {notRequired: !hearingDay.notRequired} as ActualHearingDayModel);
 
     this.hearingStore.dispatch(new fromHearingStore.UpdateHearingActuals({
       hearingId: this.id,
@@ -293,39 +140,5 @@ export class HearingActualSummaryComponent implements OnInit, OnDestroy {
     }
 
     return isValid;
-  }
-
-  public calculateEarliestHearingDate(hearingDays: ActualHearingDayModel[]): string {
-    const moments: moment.Moment[] = hearingDays.map(d => moment(d.hearingDate));
-    if (moments.length > 1) {
-      return `${moment.min(moments).format('DD MMMM YYYY')} - ${moment.max(moments).format('DD MMMM YYYY')}`;
-    } else {
-      return moment.max(moments).format(HearingDateEnum.DisplayMonth);
-    }
-  }
-
-  public getPauseStartDateTime(day) {
-    return day.pauseDateTimes && day.pauseDateTimes.length && day.pauseDateTimes[0] && day.pauseDateTimes[0].pauseStartTime
-      ? moment(day.pauseDateTimes[0].pauseStartTime).format(HearingDateEnum.DisplayTime) : null;
-  }
-
-  public getPauseEndDateTime(day) {
-    return day.pauseDateTimes && day.pauseDateTimes.length && day.pauseDateTimes[0] && day.pauseDateTimes[0].pauseStartTime
-      ? moment(day.pauseDateTimes[0].pauseEndTime).format(HearingDateEnum.DisplayTime) : null;
-  }
-
-  public isDetailsProvidedForDay(day): boolean {
-    if (this.hearingActualsMainModel.hearingActuals && this.hearingActualsMainModel.hearingActuals.actualHearingDays
-      && this.hearingActualsMainModel.hearingActuals.actualHearingDays.length > 0) {
-      const actualDay = this.hearingActualsMainModel.hearingActuals.actualHearingDays.find(d => Date.parse(d.hearingDate) === Date.parse(day.hearingDate));
-      if (actualDay) {
-        const hasActualTiming = Boolean(actualDay.hearingDate && actualDay.hearingStartTime && actualDay.hearingEndTime && actualDay.pauseDateTimes);
-        const hasActualParties = actualDay.actualDayParties.length > 0;
-
-        return hasActualTiming && hasActualParties || actualDay.notRequired;
-      }
-    }
-
-    return false;
   }
 }
